+++ drivers/net/Kconfig
@@ -373,10 +373,10 @@ config VSOCKMON
 config MHI_NET
     tristate "MHI network driver"
     depends on MHI_BUS
-    select WWAN
     ---help---
      This is the network driver for MHI bus.  It can be used with
-     QCOM based WWAN modems (like SDX55).  Say Y or M.
+     QCOM based WWAN modems for IP or QMAP/rmnet protocol (like SDX55).
+     Say Y or M.
 
 endif # NET_CORE
