--- drivers/net/mhi/Makefile
@@ -1,3 +1,0 @@
-obj-$(CONFIG_MHI_NET) += mhi_net.o
-
-mhi_net-y := net.o proto_mbim.o
